Allow Symfony 2.8 (#1)
* Dot not save composer.lock and use composer update

As a library, dependencies tests should not be locked at all.

* Add ending new line on .travis.yml (CS)

* Remove trailing whitespaces

* Allow Symfony 2.8

Symfony 2.8 IS the same as 3.0, but with the deprecated old feature.

This was done to provide an easy upgrading way between 2.x and 3.x.

Lot of people are still on 2.x version of Symfony.

Nothing has to be changed on the project codebase.

Plus, this is a LTS version. It's important to keep support on it.

Ref: http://symfony.com/blog/transition-from-symfony-2-7-to-3-0-symfony-2-8-on-its-way

* Introduce Travis matrix includes

The goal is to test the library against multiple versions of Symfony and with lowest dependencies.